src/test/java/edu/stanford/protege/metaproject/api/AccessControlObjectUUIDGeneratorTest.java
package edu.stanford.protege.metaproject.api;

import edu.stanford.protege.metaproject.api.impl.AccessControlObjectUUIDGenerator;
import org.junit.Before;
import org.junit.Test;

import static org.hamcrest.CoreMatchers.*;
import static org.hamcrest.MatcherAssert.assertThat;

/**
 * @author Rafael Gonçalves <br>
 * Stanford Center for Biomedical Informatics Research
 */
public class AccessControlObjectUUIDGeneratorTest {
    private static final String toStringHead = "AccessControlObjectUUIDGenerator";

    private AccessControlObjectUUIDGenerator gen, otherGen;

    @Before
    public void setUp() {
        gen = AccessControlObjectUUIDGenerator.getInstance();
        otherGen = AccessControlObjectUUIDGenerator.getInstance();
    }

    @Test
    public void testNotNull() {
        assertThat(gen, is(not(equalTo(null))));
    }

    @Test
    public void testGetUserId() {
        AccessControlObjectId termId = gen.getUserId();
        assertThat(termId, is(not(equalTo(null))));
        assertThat(termId.get(), is(not(equalTo(""))));
    }

    @Test
    public void testGetOperationId() {
        AccessControlObjectId termId = gen.getOperationId();
        assertThat(termId, is(not(equalTo(null))));
        assertThat(termId.get(), is(not(equalTo(""))));
    }

    @Test
    public void testGetProjectId() {
        AccessControlObjectId termId = gen.getProjectId();
        assertThat(termId, is(not(equalTo(null))));
        assertThat(termId.get(), is(not(equalTo(""))));
    }

    @Test
    public void testGetRoleId() {
        AccessControlObjectId termId = gen.getRoleId();
        assertThat(termId, is(not(equalTo(null))));
        assertThat(termId.get(), is(not(equalTo(""))));
    }

    @Test
    public void testEqualToSelf() {
        assertThat(gen, is(gen));
    }

    @Test
    public void testEquals() {
        assertThat(gen, is(otherGen));
    }

    @Test
    public void testHashCode() {
        assertThat(gen.hashCode(), is(otherGen.hashCode()));
    }

    @Test
    public void testToString() {
        assertThat(gen.toString(), startsWith(toStringHead));
    }
}